add new 0200-WebUI-Fix-EventStalling WebUI patch fixing #3385
This fixes an event stalling occurring since OCCU 3.85.5 which is caused
by a synchronous homematic() XHR execution in system.fn::UpdateUI which
is executed with the regular WebUI eventloop. This causes an event
stalling especially visible on slow connection and easily reproducible
with Firefox, due to the different main thread XHR event handling.
After removing the homematic("system.getTweak"...) call keyboard and
mouse events are not stalled anymore which should make the WebUI more
responsive again.